refactor(api): move convertApi2Food function to shared utils
